Download, clean, and transform data on presidential electoral runoff results by municipality

president_mun_vote() downloads, cleans, and transforms data on presidential elections results in the second round (runnoff elections) by municipality. The electoral results are, by default, reported as percentages.

president_mun_vote(year, perc = TRUE, ascii = FALSE,
  encoding = "windows-1252")

Arguments

year
Election year (integer). For this function, only the years 1998, 2002, 2006, 2010, and 2014 are available.

perc
Shoud the votes be reported as percentages? (Defaults to TRUE).

ascii
(logical). Should the text be transformed from Latin-1 to ASCII format?

encoding
Data original encoding (defaults to 'windows-1252'). This can be changed to avoid errors when ascii = TRUE.

Value

president_mun_vote() returns a tbl, data.frame with the following variables:

Details

This functions includes results to BR and ZZ state acronyms: the first one considers the entire country as a district; the second one considers voters living abroad as a district. For 2014 elections, results are only available for these two state acronyms.

Examples

not_run({
  df <- president_mun_vote(2002)
})