Download data on the voters' profile by vote section

voter_profile_by_section() downloads and cleans data on the voters' profile aggregated by voting section (i.e., voting stations). The function returns a data.frame where each observation corresponds to a voter profile type.

Usage

voter_profile_by_section(
  year,
  uf = "AC",
  encoding = "windows-1252",
  temp = TRUE,
  readme_pdf = FALSE
)

Arguments

year

Election year (integer). For this function, the following years are available: 2008, 2010, 2012, 2014, 2016, 2018, 2020, 2022 and 2024.

uf

Federation Unit acronym (character vector). Defaults to 'AC' (Acre).

encoding

Data original encoding (defaults to 'windows-1252'). This can be changed to avoid errors when ascii = TRUE.

temp

(logical). If TRUE, keep the temporary compressed file for future use (recommended)

readme_pdf

original readme

Value

voter_profile() returns a data.frame with the following variables:

Details

If export is set to TRUE, the downloaded data is saved as .dta and .sav files in the current directory.

Examples

if (FALSE) { # \dontrun{
df <- voter_profile_by_section(2016)
} # }